Java 1.8 compatibility changes
These changes are SPECIFIC to being able to build with Java 1.8 inside Eclipse IDE (though some changes are for all IDE's). Many applications are stuck at Java 1.8 due to some critical libraries and subcomponents of libraries that still aren't ready for Java Modules.

This checkin includes Eclipse project support files, unfortunately removed the module-info file for Java 9+ (easily recreated as it's tiny), adds the macOS ".DS-Store" files to the gitignore list, and sets the POM to use Java 1.8 vs. Java 11 as well as a slightly older version of a Maven plug-in (for compatibility due to errors at build time), and correction of the non-compliant Group ID and Artifact ID fields.

This form should NOT be merged with Origin as the main branch is rightfully targeted at Java 11 and beyond.